Updated mybatis artifacts

+++ data/src/main/java/com/itt/data/dao/EventsDao.java
@@ -1,12 +1,7 @@
-/*
- * Copyright (c) 2020. Robert Mayore
- */
-
 package com.itt.data.dao;
 
 import com.itt.data.model.Event;
 import com.itt.data.model.EventExample;
-import java.math.BigDecimal;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
@@ -15,7 +10,7 @@ public interface EventsDao {
 
     int deleteByExample(EventExample example);
 
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Event record);
 
@@ -23,7 +18,7 @@ public interface EventsDao {
 
     List<Event> selectByExample(EventExample example);
 
-    Event selectByPrimaryKey(BigDecimal id);
+    Event selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Event record, @Param("example") EventExample example);
